Add files via upload
This is the most important step for you now. You are currently writing the commit message in the text box.

First Line (Summary): The first line should be a short, descriptive summary of the changes. Something like: Add initial files for Snake Ladder game or Initial commit of basic project structure.

Body (Details): If you need more detail, you can press Enter and write a longer description on the following lines. It looks like you've started one with "This is the basic project and it contains...". You can finish that thought.